requirements.txt
apache-airflow-providers-postgres
pandas
numpy
python-dotenv
